refactor: 메인 페이지 컴포넌트를 한 곳에 모아 barrel export 적용
