--- src/components/Ending/index.ts
@@ -1,1 +1,0 @@
-export { Ending } from './Ending';
