--- src/components/Intro/index.ts
@@ -1,1 +1,0 @@
-export { Intro } from './Intro';
