+++ src/components/Main/Ending.tsx
