+++ src/components/Main/index.tsx
@@ -1,2 +1,7 @@
+export { Ending } from './Ending';
+export { Intro } from './Intro';
+export { Journey } from './Journey';
+export { RecruitEntrance } from './RecruitEntrance';
 export { RecruitTextSection } from './RecruitTextSection';
 export { SignImageSection } from './SignImageSection';
+export { Subscribe } from './Subscribe';
