--- src/components/Subscribe/index.ts
@@ -1,1 +1,0 @@
-export { Subscribe } from './Subscribe';
